\name{qsbralks}
\alias{qsbralks}
\docType{data}
\title{
qsbralks
}
\description{
 The dataset consists of 21 variables and 13 observations with a continuous response. This dataset is taken from http://www.openml.org/d/431.
}
\usage{data("qsbralks")}
\format{
  A data frame with 13 observations and 22 variables including the response

  }


